Removed Modelling Portion, Added Rolling Sum Plot
I have the goal of creating a few models via some machine learning
techniques so that I can implement what I've learned so far from
my reading of technical papers and schooling, however that is a
rather large task and deserves a separate notebook or script, so
I'll be saving that for later.

I've added in a plot showing the rolling sum of rainfall levels vs
reservoir levels that I think gets the point across that there's
a time factor that would need to be accounted for my any model
(hence my want to use a recurrent neural net for modeling).